--- src/mocks/handlers/RegisterUserForm/queryIsAliasUnique.ts
@@ -1,16 +1,0 @@
-import { graphql } from "msw";
-
-import { RegisterUserIsAliasUniqueDocument } from "~/mocks/codegen";
-import { c } from "~/mocks/constraints";
-
-const aliases = Object.keys(c.users);
-
-export const queryIsAliasUnique = graphql.query(
-  RegisterUserIsAliasUniqueDocument,
-  (req, res, ctx) => {
-    return res(ctx.data({
-      __typename: "Query",
-      findUser: { __typename: "FindUserPayload", user: { __typename: "User", id: "id" } },
-    }));
-  },
-);
