--- src/mocks/handlers/index.ts
@@ -1,22 +1,0 @@
-// global
-export { queryFetchViewer } from "./queryFetchViewer";
-
-// create henken form
-export { mutationCreateHenken as mutationCreateHenkenFormCreateHenken } from "./CreateHenkenForm/mutationCreateHenken";
-export { querySearchContent as queryCreateHenkenFormSearchContent } from "./CreateHenkenForm/querySearchContent";
-export { querySearchUser as queryCreateHenkenFormSearchUser } from "./CreateHenkenForm/querySearchUser";
-
-// global nav
-export { queryNotifications as queryGlobalNavNotifications } from "./GlobalNav/queryNotifications";
-
-// register form
-export { mutationRegisterUser as mutationRegisterUserFormRegisterUser } from "./RegisterUserForm/mutationRegisterUser";
-export { queryIsAliasUnique as queryRegisterUserFormIsAliasUnique } from "./RegisterUserForm/queryIsAliasUnique";
-
-// henken page
-export { queryAllHenkenPages, queryHenkenPage } from "./HenkenPage";
-
-// user page
-export { queryAllUserPages } from "./UserPage/queryAllUserPages";
-export { querySendHenkensUserPage } from "./UserPage/querySendHenkensUserPage";
-export { queryUserPage } from "./UserPage/queryUserPage";
