--- src/mocks/next/server.ts
@@ -1,5 +1,0 @@
-import { setupServer } from "msw/node";
-
-import { handlers } from "./handlers";
-
-export const mockServer = setupServer(...handlers);
