--- src/mocks/test/server.ts
@@ -1,3 +1,0 @@
-import { setupServer } from "msw/node";
-
-export const mockServer = setupServer();
